requirements.txt
flask>=2.0.0
flask-sqlalchemy>=3.0.0
flask-caching>=2.0.0
apscheduler>=3.10.0
selenium>=4.11.2
webdriver_manager>=4.0.2
tenacity>=8.0.0